+++ tests/bundle.js
@@ -1,3 +1,3 @@
-const { ParentSchema, ExtraMetaSchema } = require('./bundle-schematics');
+const { parentSchema, extraMetaSchema } = require('./bundle-schematics');
 // Lets include some other stuff too
 const PropTypes = require('prop-types');