+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # 5.0.0
+* Integrated `optimal` for options validation.
 * Moved flow definitions to flow-typed/ folder.
-* Renamed `shapeshifter.js.flow` to `shapeshifter.flow.js`.
 
 # 4.2.1
 * Fixed an issue with the lib/ build.